fix notifications alert handling in new testing setup
we used to rely on the onboarding test running first and approving alerts for us, but that no longer is the case.
